Moves NamedEvent native code to its own files
